client/src/Header/__test__/Header.test.tsx
import React from 'react';
import { mount } from 'enzyme';
import Header, { HeaderProps } from '../';

const defaultProps: HeaderProps = undefined;
const mountComponent = (customProps: any = {}) => {
  const props = { ...defaultProps, ...customProps };
  const wrapper = mount(<Header {...props} />);

  return { props, wrapper };
};

describe('Header', () => {
  describe('rendering', () => {
    it('should render component without any crash', () => {
      const { wrapper } = mountComponent();
      expect(wrapper).toHaveLength(1);
    });

    it('should be match with snapshot', () => {
      const { wrapper } = mountComponent();
      expect(wrapper).toMatchSnapshot();
    });
  });

  describe('props', () => {
    describe('default', () => {
      it('should have expected default props', () => {
        expect(JSON.stringify(defaultProps)).toBe(
          JSON.stringify(Header.defaultProps)
        );
      });
    });
  });
});
